Add `mongo-express` container for DB interaction
MongoDB Compass was initially considered for interacting with the
database, but it requires additional setup and installation on the
developer's side, which we can't control.

To streamline the process, we opted for `mongo-express`, a web-based UI
that can be integrated into the project’s Docker setup. This solution
simplifies database interaction without requiring external tools to be
installed.

`mongo-express` is well-maintained (last updated in Dec 2024), provides an intuitive interface, and has an official Docker image.

To access the UI, visit: http://localhost:8081 with the following credentials:

 Username: admin
 Password: pass

mongo-express repo: https://github.com/mongo-express/mongo-express