Suppress outdated JSX transform warning in tests
A warning about using an outdated JSX transform appeared during test
runs. After confirming that it didn't cause any issues or test failures,
we investigated but couldn't find a direct fix.

To avoid noise in test output, we added a suppression in the test setup
to filter out this specific warning.

console.warn
    Your app (or one of its dependencies) is using an outdated JSX transform. Update to the modern JSX transform for faster performance: https://react.dev/link/new-jsx-transform

       5 | describe('Home', () => {
       6 |   it('renders a heading', () => {
 7 |     render(<Counter />)
         |            ^
       8 |
       9 |     const heading = screen.getByRole('heading', { level: 1 })
      10 |

      at Object.<anonymous>.process.env.NODE_ENV.exports.createElement (node_modules/react/cjs/react.development.js:991:19)
      at Object.__jsx (src/tests/components/Counter.test.tsx:7:12)