Add util funcs to handle xp related calculations
This commit introduces a function to create the level map, to determine
how much xp is needed for someone to progress to the next level.

It also includes a function to calculate how much progress has been made
towards the next level. This is relevant to the Profile Page and other components.